Refactor table component imports and alignment
The table components in the DesignSystem have been refactored for improved readability and maintainability. This refactor includes adjusting the alignment of elements for a more consistent code style, and updating the way components and types are imported to follow best practices in TypeScript and Vue. Additionally, the ITableColumn interface has been renamed to ITableColumnData for better clarity.