Refactor table sorting and improve layout
The table sorting functionality has been optimized by introducing a computed class in the TableThSorting.vue to manage the opacity changes according to the state. The layout of the index page has been restructured to include a sorted and striped table inside a div container. Minimal formatting changes were made on DsIcon.vue for clarity and consistency.