Refactor data structure and adjust table styling
Improved the structure of the data array passed to the DsTable component in pages/index.vue, making it easier to understand and manipulate. Additionally, a minor adjustment was made to the styling of the DsTable component to better handle scenarios when the data array is empty, preventing table distortions that previously occurred.