Update data display in table
The data and columns displayed in the table have been updated: Data has been changed to display information about benefits and procedures, and columns have been modified to include information about the name, date, and category of the items. The action buttons have been realigned to the right and the table now supports clickable links.