Add cancel button functionality to CategoryForm
Added an `emit` function to the 'Cancelar' (Cancel) button on the 'CategoryForm' Vue component in CategoryForm.vue file. Also implemented a handler for this action in index.vue, which changes the view state from 'form' to 'table' when the 'Cancelar' button is clicked, switching the view back to the categories list.