Added Pinia state management and refactored app.vue
In this update, Pinia was added as a new dependency and incorporated into the Nuxt configuration. This also involved the creation of a new file, builderStore.ts, to define and export the useCounterStore. Corresponding changes were applied to app.vue, replacing Axios and postData function with Pinia's store functions to handle increment and changing the name.