Implement dynamic side menu switching
The commit introduces dynamic switching between the default side menu and the builder menu in the application. The side menu type can now be controlled via a state in the CounterStore. Conditions to display the default or the builder menu have been added to default.vue. The functionality has been simplified in app.vue by removing unnecessary increment and name change actions, and adding menu change action. Furthermore, an extra option for the builder menu has been added to the verticalMenu data.