Removed navigation links and moved menu change to middleware
The navigation links in DsUnidosLayout.vue have been commented out as part of a redesign. The DsButton to change the menu has been removed from app.vue as this functionality has been moved to the new middleware, default.global.ts. The middleware now changes the side menu type based on the route, eliminating the need for a button to switch between menu types.