Update builder functionality in Drag and Drop
In this update, the drag and drop functionality within the builder has been enhanced. An addItemToForm function has been added in builderStore which adds the current item to the form upon dropping. Additionally, the representation of builder items in DraggableArea was updated to display item.name instead of item.type. CurrentDragItem display in BuilderMenu was also removed.