Update layout and styling in several Vue components
Enhanced the user interface by updating the styling and layout in 'default.vue', 'index.vue', and 'DraggableArea.vue'. This includes adding padding and width to certain elements, changing the flex direction, and adding a new 'Tools' section in 'index.vue'. This should improve the overall appearance and user experience of the application.