Update PropertyPanel and InputPanel components
A component key ref has been added to PropertyPanel.vue to handle deep watch of currentEditItem. It increments on every change in currentEditItem. This key is then passed to `<InputPanel/>`, `<SelectPanel/>`, and `<TextAreaPanel/>` components. In InputPanel.vue, the initial state of labelValue' ref is updated to reflect currentEditItem's label property.