Implement code view and edit functionality in Builder page
A new feature has been added to transition between code and edit views in the Builder page. A "CodeArea" component has been introduced and two new methods have been added to handle the switching of views in the Builder. The "ToolPanel" component is now capable of emitting events corresponding to both modes.