Add clear store feature to the builder
This commit adds the functionality to clear the entire form in the builder. The new clear store function was created and added to the builder store where it empties the array holding the current form items. This functionality was then integrated with the user interface via the ToolPanel component, allowing the clear store action to be triggered by a user's interaction.