Optimize item update and fix store clearing
This commit optimizes the updateItemInForm function in the builderStore by finding the index of the item that needs to be updated, instead of directly using the item's id. Additionally, it fixes the clearStore function by properly clearing the builderItem, currentEditItem and currentDragItem values. A debugging line is also added in the DraggableArea.vue to log the builderItems data structure.