Implement conditional draggability for DsConfirmationButton
Updated the BuilderMenu and item components to conditionally allow the DsConfirmationButton to be dragged. This change was implemented by adding a 'draggable' property to item component, and a computed property 'isDsConfirmationButtonPresent' in BuilderMenu to check if the DsConfirmationButton is already present, to prevent multiple instances of the same button from being created.