Implement drag functionality for version history and modify builder store
The commit implements drag functionality for version history in the VersionArea component. Properties in the IItemBuilder interface are now optional. The builderStore now has the ability to handle a group of items instead of just one. It also removes the generateId function since IDs now come with the data itself. The mock data for different versions has also been extended.