Simplify ToolPanel event binding in builder index
In the builder index,this commit simplifies the way events are passed to the ToolPanel component. Instead of individually binding each event (@clear, @code, @edit, @save), we've updated the code to use Vue's `v-on` shorthand to bind all events from the `handle` object at once. This not only reduces the lines of code but also increases readability and ease of maintenance.