Update draggable properties in DraggableArea component
The draggable attribute in the DraggableArea component of the builder-components has been modified. Not only has it been set to false for the main div, but the draggable attribute has also been removed from the builder items. A prevent action has been added to the dragstart event to ensure a smooth user experience.