Update UI components in builder
Updated the ValidatePanel and DraggableArea components in the builder. In the ValidatePanel component, replaced the static display of current item name with an input field to allow direct editing. In the DraggableArea component, updated the property reference from item.name to item.type to correctly bind Vue components. These changes improve user interaction and functionality in the UI builder.