Update property references from 'name' to 'type'
The usage of 'name' property to identify the types of items in 'builderStore' and 'PropertyPanel' components has been replaced with 'type'. This change provides a more accurate description of the item type. Replacing 'name' with 'type' will ensure that the item's identity is not mistaken with its type, making the code more intuitive and maintainable.