Update form validation and default values
Updated form input validation properties with custom rules and configured default field validation type. In addition, added a mechanism in the ViewArea.vue component to create and update validation rules based on the validation properties of each form item. Consequently, the field validation type for newly created items is set to a default type in the builderStore.ts store.