Add functionality to view form data in beautified format
A functionality was added that allows viewing the form data in a beautified format. This was achieved by using Prism for syntax highlighting, and the js-beautify library for beautifying the JSON data. In conjunction with this, a watcher was implemented to track changes in form values and automatically update them.