Implement collapsible feature for property panel
This commit implements functionality to collapse and expand sections of the property panel. Along with this, it also updates class types in the typography component and includes invocations of the accordion component in various other components. The builder store has been updated to manage the state of the property panel collapse.