Add Info Panel to builder with modal functionality
A new functional component called InfoPanel has been implemented with conditional rendering based on modalType. InfoPanel includes an explanatory text in Spanish about its use. It is made visible by changing the modalType to 'infoPanel' which checks if there are builderItems. If none are present, the modal won't be displayed. The 'infoPanel' option has been added to the IModalType to accommodate these changes.